Reworked command execution and logging into a single function and implemented a dummy mode (no changes done, only shows commands that would have been executed).
Changed logging to go to stderr.

Implemented persistent storage of LU GUID and LUN into a user-property in the
zvol to allow for zvol renaming. The initial GUID is still generated from an
MD5 of the initial zvol name.

Implemented parallel execution of delete-lu, delete-target and create-lu to
speed up import/export of pools with many zvols.