Fixes for PCS
If the DEBUG variable is not set, line 35 fails with incorrect comparison. It has to be either '0' or '1' (or any integer value)
zpool_is_imported failure (because it is not imported initially) results in OCF_STD_ERR which results in error messages of the resource in PCS, despite loading and mounting.